Added FB messenger module to testbot
Added FB messenger module to our bot using the following instructions:
https://www.youtube.com/watch?v=HTpUmDz9kRY

Added the bot via the bot dashboard, accessible at
http://localhost:3000 after running `bp start` from within the `/testbot`
folder.

Module appears in the dash and clicking on it brings up its config page.

Next step is to configure the module and connect it to an FB page
with the appropriate developer/API keys.

Responds to Issue #2.