Started deployment checklist for testbot
Began deployment checklist for testbot.

Includes instructions on how to run testbot on localhost:3000 and
how to use ngrok to connect the bot to Facebook using an ngrok
webserver that takes incoming connections to converse with the FB
API.

Next step is to deploy the bot to Heroku and writeup the checklist
section for deploying and running the bot on Heroku.

Responds to Issue #11.